Address tracing review feedback and update DataShard expectations
