+++ ydb/core/kqp/common/compilation/result.h
@@ -1,6 +1,5 @@
 #pragma once
 #include <memory>
-#include <ydb/core/kqp/common/simple/query_stats.h>
 #include <util/datetime/base.h>
 #include <util/digest/multi.h>
 #include <ydb/core/kqp/common/simple/query_ast.h>
@@ -71,7 +70,11 @@ struct TKqpCompileResult {
     TMaybe<NJson::TJsonValue> CompileMeta; // consist of parameters of query
 };
 
-
+struct TKqpStatsCompile {
+    bool FromCache = false;
+    ui64 DurationUs = 0;
+    ui64 CpuTimeUs = 0;
+};
 } // namespace NKikimr::NKqp
 
 template<>
