+++ ydb/core/kqp/session_actor/kqp_query_stats.h
@@ -1,12 +1,30 @@
 #pragma  once
 
-#include <ydb/core/kqp/common/simple/query_stats.h>
+#include <ydb/core/kqp/common/compilation/result.h>
 #include <ydb/core/base/defs.h>
 #include <ydb/core/protos/kqp_stats.pb.h>
-#include <ydb/core/protos/kqp.pb.h>
 
 namespace NKikimr::NKqp {
 
+struct TKqpQueryStats {
+    ui64 DurationUs = 0;
+    ui64 QueuedTimeUs = 0;
+    std::optional<TKqpStatsCompile> Compilation;
+
+    ui64 WorkerCpuTimeUs = 0;
+    ui64 ReadSetsCount = 0;
+    ui64 MaxShardProgramSize = 0;
+    ui64 MaxShardReplySize = 0;
+    ui64 LocksBrokenAsBreaker = 0;
+    ui64 LocksBrokenAsVictim = 0;
+
+    TVector<NYql::NDqProto::TDqExecutionStats> Executions;
+
+    const TVector<NYql::NDqProto::TDqExecutionStats>& GetExecutions() const;
+    ui64 GetWorkerCpuTimeUs() const;
+
+    NKqpProto::TKqpStatsQuery ToProto() const;
+};
 
 
 void CollectQueryStats(const TActorContext& ctx, const NKqpProto::TKqpStatsQuery* queryStats,
