+++ ydb/core/kqp/tracing/test_util/ya.make
@@ -1,0 +1,12 @@
+LIBRARY()
+
+SRCS(
+    kqp_trace_test_helpers.h
+)
+
+PEERDIR(
+    library/cpp/testing/unittest
+    ydb/library/actors/wilson/test_util
+)
+
+END()
