+++ ydb/core/kqp/tracing/ya.make
@@ -23,3 +23,7 @@ PEERDIR(
 )
 
 END()
+
+RECURSE_FOR_TESTS(
+    ut
+)
