+++ ydb/core/protos/kqp.proto
@@ -341,7 +341,6 @@ message TEvQueryResponse {
     reserved 5; // (deprecated) ForcedNewEngine
     optional uint64 ConsumedRu = 6;
     optional bool WorkerIsClosing = 7 [default = false];
-    reserved 8; // UserFacingTrace
     optional ERejectionStage RejectionStage = 9;
 }
