+++ ydb/library/actors/wilson/test_util/fake_wilson_uploader.h
@@ -4,10 +4,25 @@
 #include <ydb/library/actors/core/hfunc.h>
 #include <ydb/library/actors/wilson/wilson_uploader.h>
 
+#include <library/cpp/threading/future/future.h>
+
+#include <util/stream/str.h>
+#include <util/system/mutex.h>
+
+#include <functional>
+#include <optional>
+#include <queue>
+#include <set>
+#include <string>
+#include <unordered_map>
+#include <vector>
+
 namespace NWilson {
 
-    class TFakeWilsonUploader : public NActors::TActorBootstrapped<TFakeWilsonUploader> {
-        public:
+    class TTraceSnapshot {
+    public:
+        using TOtelSpan = opentelemetry::proto::trace::v1::Span;
+
         class Span {
         public:
             Span(TString name, TString parentSpanId, ui64 startTime) : Name(name), ParentSpanId(parentSpanId), StartTime(startTime) {}
@@ -109,12 +124,7 @@ namespace NWilson {
         };
 
     public:
-        void Bootstrap() {
-            Become(&TThis::StateFunc);
-        }
-
-        void Handle(NWilson::TEvWilson::TPtr ev) {
-            TOtelSpan& span = ev->Get()->Span;
+        void AddSpan(const TOtelSpan& span) {
             TOtelSpan spanCopy;
             spanCopy.CopyFrom(span);
             Spans.push_back(std::move(spanCopy));
@@ -159,13 +169,7 @@ namespace NWilson {
             Traces.clear();
         }
 
-        STRICT_STFUNC(StateFunc,
-            hFunc(NWilson::TEvWilson, Handle);
-        );
-
     public:
-        using TOtelSpan = opentelemetry::proto::trace::v1::Span;
-
         std::unordered_map<TString, Trace> Traces;
         std::vector<TOtelSpan> Spans;
 
@@ -178,4 +182,32 @@ namespace NWilson {
         }
     };
 
-} // NWilson
+    class TFakeWilsonUploader : public NActors::TActorBootstrapped<TFakeWilsonUploader>, public TTraceSnapshot {
+    public:
+        struct TEvGetSnapshot : NActors::TEventLocal<TEvGetSnapshot, EventSpaceBegin(NActors::TEvents::ES_PRIVATE) + 1> {
+            explicit TEvGetSnapshot(NThreading::TPromise<std::vector<TOtelSpan>> promise)
+                : Promise(std::move(promise))
+            {}
+
+            NThreading::TPromise<std::vector<TOtelSpan>> Promise;
+        };
+
+        void Bootstrap() {
+            Become(&TThis::StateFunc);
+        }
+
+        void Handle(NWilson::TEvWilson::TPtr ev) {
+            AddSpan(ev->Get()->Span);
+        }
+
+        void Handle(TEvGetSnapshot::TPtr ev) {
+            ev->Get()->Promise.SetValue(Spans);
+        }
+
+        STRICT_STFUNC(StateFunc,
+            hFunc(NWilson::TEvWilson, Handle);
+            hFunc(TEvGetSnapshot, Handle);
+        );
+    };
+
+} // namespace NWilson
