+++ ydb/library/actors/wilson/test_util/ya.make
@@ -4,4 +4,8 @@ SRCS(
     fake_wilson_uploader.h
 )
 
+PEERDIR(
+    library/cpp/threading/future
+)
+
 END()
